[CMS] Update, copy to clipboard file url [CMS] Possible leave the same name while adding new file (change)

+++ !cms/php/run/file/init.php
@@ -10,9 +10,12 @@ if($newFile != '') {
 
     if($fileName = $addition->setFileName($newFile, $permitted)) {
 
+        if($p_choose)
+            $fileName = $eventData['current'];
+
         if($addition->addFile($tmpName, $eventData['path'] . $fileName)) {
 
-            if (isset($eventData['current'])) {
+            if (isset($eventData['current']) and !$p_choose) {
 
                 if(!$addition->removeFile($eventData['path'] . $eventData['current']))
                     $alert0 = $translation['validation']['delete-file-fail'];
